refactor: remove slice (#196)
`Slice` and `SliceMut` are introduce to solve async io lifetime
problem in unsafe way. Remove slices to make code easier.

+++ foyer-storage/src/buffer.rs
@@ -127,10 +127,7 @@ where
         debug_assert!(self.offset + buffer.len() <= self.device.region_size());
 
         // flush and clear buffer
-        let (res, mut buffer) = self
-            .device
-            .write(buffer, .., region, self.offset as u64)
-            .await;
+        let (res, mut buffer) = self.device.write(buffer, .., region, self.offset).await;
         buffer.clear();
         self.buffer = Some(buffer);
         res?;
